refactoring: move get_admin_url() to mixin
